discard overkill refactoring
